% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_util.R
\name{.prep_pattern}
\alias{.prep_pattern}
\title{Utility function to collapse a character vector into a Regex pattern. The
vector elements will be surrounded by word boundaries, and combined with
logical ORs. The whole thing is wrapped in parantheses to catch it as a
group.}
\usage{
.prep_pattern(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
The elements of \code{x} surrounded by Regex word boundaries,
collapsed with logical ors, and surrounded by parantheses:
(\\bx[1]\\b|\\bx[2]\\b|...).
}
\description{
Utility function to collapse a character vector into a Regex pattern. The
vector elements will be surrounded by word boundaries, and combined with
logical ORs. The whole thing is wrapped in parantheses to catch it as a
group.
}
\examples{
.prep_pattern(c("foo", "bar"))
}
