% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{.remove_diacritics}
\alias{.remove_diacritics}
\title{Utility functions for string processing.}
\usage{
.remove_diacritics(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
\code{string} with diacritical letters replaced by their ASCII
  versions.
}
\description{
Replace diacritical letters(é, ç, ...) with their "plain" versions. This
function can only handle diacritical letters from latin alphabets. Elements
in \code{string} containinig non-latin letters (e.g. cyrillic), will be
replaced by \code{NA}.
}
\details{
Reference: https://stackoverflow.com/a/20495866/13542638
}
\examples{
thoremisc:::.remove_diacritics("Åll thëşé fūñny leŧters wîll be nørmalised.")
}
