% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{.remove_special_chars}
\alias{.remove_special_chars}
\title{Utility functions for string processing.}
\usage{
.remove_special_chars(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
\code{string} with special Unicode characters replaced by a
  whitespace.
}
\description{
Replace any characters that do not belong to Regex classes \\w or \\d, or are
a literal whitespace, by a single whitespace. The function preserves German
Umlaute and diacritical letters.
}
\details{
Elaboration on the Regex classes:
https://stackoverflow.com/a/2998550/13542638.
}
\examples{
thoremisc:::.remove_special_chars("This will be modified: hello-world.")
}
