% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{.replace_umlaute}
\alias{.replace_umlaute}
\title{Utility functions for string processing.}
\usage{
.replace_umlaute(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
\code{string} with German Umlaute replaced.
}
\description{
Replace German Umlaute by their ASCII representations: "ä"->"ae", "ö"->"oe",
and "ü"->"ue". "ß" is diacritical, see \code{.remove_diacritics()}.
}
\examples{
thoremisc:::.replace_umlaute("Äh, trörö in Überlingen, nicht auf dem Darß.")
}
