% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_acro.R
\name{latex_acro}
\alias{latex_acro}
\title{Wrap abbreviations in text in a LaTeX command.}
\usage{
latex_acro(string, acros, command = "acs*")
}
\arguments{
\item{string}{A character vector.}

\item{acros}{A character vector of abbreviations to wrap.}

\item{command}{A character (without the preceding backslash).}
}
\value{
\code{string} with all \code{acros} wrapped in \code{command}.
}
\description{
Wrap abbreviations in text in a LaTeX command.
}
\examples{
latex_acro("This package is not on CRAN.", "CRAN")
}
