% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_captions.R
\name{latex_add_caption}
\alias{latex_add_caption}
\title{Add a caption to a stargazer::stargazer() output.}
\usage{
latex_add_caption(string, caption, to_bottom = TRUE)
}
\arguments{
\item{string}{An atomic character produced by stargazer::stargazer().}

\item{caption}{An atomic character to be added as the caption.}

\item{to_bottom}{Boolean, should the caption moved to the bottom?}
}
\value{
string with an added caption
}
\description{
Add a caption to a stargazer::stargazer() output.
}
\examples{
latex_add_caption(stargazer::stargazer(model), "This is going to be the
caption", TRUE)
}
