% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_captions.R
\name{latex_caption_to_bottom}
\alias{latex_caption_to_bottom}
\title{Move the caption to the bottom of the table produced by
\code{stargazer::stargazer()}.}
\usage{
latex_caption_to_bottom(string, to_console = FALSE)
}
\arguments{
\item{string}{An atomic character produced by \code{stargazer::stargazer()}.}

\item{to_console}{Boolean, should the output be returned, or \code{cat()} to
the console (and invisibly returned)?}
}
\value{
\code{string} with the caption moved to the bottom.
}
\description{
Credit to https://stackoverflow.com/a/42681368
}
\examples{
df <- data.frame(a = rnorm(5), b = rnorm(5))
sg <- stargazer::stargazer(df)
latex_caption_to_bottom(sg)
}
