% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_greeks.R
\name{latex_greeks}
\alias{latex_greeks}
\title{Precede greek letters with backslashes to render them in LaTeX.}
\usage{
latex_greeks(string, alphabet = NULL, append = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{alphabet}{A character vector of strings to be preceded by backslashes.}

\item{append}{Boolean, should the supplied \code{alphabet} be appended to the
implemented greek letters.}
}
\value{
\code{string} with all greek letters, and what is supplied in
\code{alphabet} preceded by backslashes.
}
\description{
Precede greek letters or whatever supplied in \code{alphabet} with
backslashes, such that when \code{cat()} is applied, the output can be
copied into a \code{.tex} document and be rendered.
}
\examples{
latex_greeks("alpha, beta, and foo will be modified; bar won't", "foo", TRUE)
}
