% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_dummy_dates.R
\name{pan_dummy_dates}
\alias{pan_dummy_dates}
\title{Create dummy panel from start and end dates.}
\usage{
pan_dummy_dates(
  x,
  g,
  col_start,
  col_end,
  freq = c("m", "q", "y"),
  t_name,
  d_name,
  alt_end = NULL
)
}
\arguments{
\item{x}{A \code{data.frame} with events' start and end dates per row.}

\item{g}{An atomic character, name of the panel group variable in \code{x}.}

\item{col_start}{Atomic character, name of the column of start dates in
\code{x}. As of now, must be of type \code{zoo::yearmon}.}

\item{col_end}{Atomic character, name of the column of end dates in \code{x}.
As of now, must be of type \code{zoo::yearmon}.}

\item{freq}{Atomic character, the frequency of the resulting panel. One of
\code{m} (monthly), \code{q} (quarterly), \code{y} (yearly).}

\item{t_name}{An atomic character, name of the panel time variable in the
resulting \code{data.frame}.}

\item{d_name}{An atomic character, name of the panel group variable in the
resulting \code{data.frame}.}

\item{alt_end}{Atomic numeric, alternative last period for resulting panel.
If \code{NULL}, the resulting panel extends the last date supplied in
\code{col_end}.}
}
\value{
A \code{data.frame} or \code{data.table} depending on the input.
}
\description{
Create a dummy panel over time from data on events with start and end dates.
Codes any month, quarter, or year (depending on \code{freq}) during which an
event takes place is dummy coded as \code{1}.
}
\details{
Currently supported are yearly, quaterly, and monthly data.
}
\examples{
x <- data.frame(
  grp = rep(c("A", "B"), each = 3),
  start_month = zoo::as.yearmon(
    c("2000-01-01", "2002-05-01", "2003-12-01", "2000-06-01", "2003-01-01",
      "2006-01-01")
    ),
  end_month = zoo::as.yearmon(
    c("2000-05-01", "2003-02-01", "2004-03-01", "2001-06-01", "2005-07-01",
      "2008-01-01")
    )
)

pdt <- pan_dummy_dates(
  x, "grp", "start_month", "end_month", "m", "yearm", "dummy",
  zoo::as.yearmon("2008-12-01")
)
}
