% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pan_labeller_country.R
\name{pan_labeller_country}
\alias{pan_labeller_country}
\title{Labeller function for \code{ggplot2} \code{facet}s.}
\usage{
pan_labeller_country(
  x,
  origin = "iso3c",
  destination = "country.name",
  both = TRUE
)
}
\arguments{
\item{x}{A character vector.}

\item{origin}{An atomic character, see \code{?countrycode::codelist}.}

\item{destination}{An atomic character, see \code{?countrycode::codelist}.}

\item{both}{Boolean, should origin and destination be returned, or only
destination?}
}
\value{
A character vector.
}
\description{
Takes a character vector of some origin format given in
\code{?countrycode::codelist} and returns either a character vector of the
format [origin] - [destination], or a character vector of only the
destination format.
}
\examples{
pan_labeller_country(c("AUS", "TWN"))
}
