% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_clean.R
\name{string_clean}
\alias{string_clean}
\title{Clean letters in a string.}
\usage{
string_clean(
  string,
  lower = TRUE,
  spec_chars = TRUE,
  umlaute = TRUE,
  diacritics = TRUE
)
}
\arguments{
\item{string}{A character vector.}

\item{lower}{Boolean, convert \code{string} to lowercase.}

\item{spec_chars}{Boolean, remove non-letters.}

\item{umlaute}{Boolean, replace German Umlaute.}

\item{diacritics}{Boolean, transliterate diacritical letters.}
}
\value{
\code{string}, modified.
}
\description{
Transform \code{string} into a simple lowercase ASCII representation. Options
to convert \code{string} to lowercase, remove non-letter characters, replace
German Umlaute, and transliterate diacritical letters (á, å, ê, ...).
}
\details{
The function removes full stops, i.e. breaks sentences.
}
\examples{
string_clean("Thîs sŧriñg will bé å løt simplêr. Köln is not in M-V.")
}
