% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_redund_ws.R
\name{string_redund_ws}
\alias{string_redund_ws}
\title{Remove redundant whitespaces from a string.}
\usage{
string_redund_ws(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
\code{string} with redundant whitespaces removed.
}
\description{
Replace consecutive whitespaces by a single one, and trim leading and
trailing whitespace.
}
\examples{
string_redund_ws(" This  string   will be shorter.   ")
}
