% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_stem_group.R
\name{string_stem_group}
\alias{string_stem_group}
\title{Stem all words in a group.}
\usage{
string_stem_group(string, language)
}
\arguments{
\item{string}{A character vector containing strings of 1 or more words.}

\item{language}{Passed to \code{SnowballC::wordStem()}, see
\code{SnowballC::getStemLanguages()}.}
}
\value{
\code{string}, with all individual words stemmed, collapsed back into
  the original groups of words
}
\description{
Take a character vector of groups of words, and stem all words in each group.
The results are collapsed back into the original grouping, with words
separated by a single whitespace. This function wraps
\code{SnowballC::wordStem()} which can only handle single words as input.
}
\details{
Designed to take cleaned strings (see \code{string_clean()}) as input.
}
\examples{
string_stem_group(c("The coolest words", "Will all be stemmed"), "english")
}
