% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_util.R
\name{string_utility}
\alias{string_utility}
\alias{.remove_special_chars}
\alias{.replace_umlaute}
\alias{.remove_diacritics}
\title{String processing utility functions.}
\usage{
.remove_special_chars(string)

.replace_umlaute(string)

.remove_diacritics(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
\code{.remove_special_chars} returns \code{string} with non-letter
  Unicode characters replaced by a whitespace.

  \code{.replace_umlaute} returns \code{string} with any German Umlaute
  replaced.

  \code{.remove_diacritics} returns \code{string} with diacritical letters
  replaced by their ASCII versions.
}
\description{
Short utility functions to clean certain characteristics of strings. These
are combined in \code{\link{string_clean}}.
}
\details{
Replace any characters that do not belong to Regex classes \\w or \\d, or are
a literal whitespace, by a single whitespace. The function preserves German
Umlaute and diacritical letters.

Elaboration on the Regex classes:
https://stackoverflow.com/a/2998550/13542638.

Replace German Umlaute by their ASCII representations: "ä"->"ae", "ö"->"oe",
and "ü"->"ue". "ß" is diacritical and handled by \code{.remove_diacritics}.

Replace diacritical letters(é, ç, ...) with their "plain" versions. This
function can only handle diacritical letters from latin-based alphabets.
Elements in \code{string} containinig non-latin letters (e.g. cyrillic), will
be replaced by \code{NA}.

Reference: https://stackoverflow.com/a/20495866/13542638
}
\examples{
thoremisc:::.remove_special_chars("This will be modified: hello-world.")
thoremisc:::.replace_umlaute("Äh, trörö in Überlingen, nicht auf dem Darß.")
thoremisc:::.remove_diacritics("Åll thëşé fūñny leŧters wîll be nørmalised.")
}
\seealso{
\code{\link{string_clean}} and \code{\link{string_redund_ws}}
}
