% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_event_count.R
\name{ts_event_count}
\alias{ts_event_count}
\title{Count and ennumerate of events in a dummy series.}
\usage{
ts_event_count(x)
}
\arguments{
\item{x}{A numeric vector of \code{0}s and \code{1}s.}
}
\value{
A numeric vector containing \code{NA}s, and the count of the event
corresponding to the dummy series.
}
\description{
Given a numeric vector with dummies for events, ennumerate the events. Return
a vector of \code{length(x)} filled with \code{NA}s, and the count, where
\code{x} equals \code{1}.
}
\examples{
ts_event_count(c(0, 0, 1, 1, 0, 1, 0))
}
