% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_event_dist.R
\name{ts_event_dist}
\alias{ts_event_dist}
\title{Return a information on the occurence of a dummy}
\usage{
ts_event_dist(x)
}
\arguments{
\item{x}{A numeric vector of `0`s and `1`s.}
}
\value{
A data.frame with two numeric columns.
}
\description{
Returns a data.frame of `length(x)` rows with a column of the number of rows
or steps until the next time the dummy series equals 1, and a columns of the
number of rows or steps since the last time the dummy series equaled 1.
}
\examples{
ts_event_dist(c(0, 0, 0, 1, 0, 1, 0))
}
