% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_pad_contig.R
\name{ts_pad_contig}
\alias{ts_pad_contig}
\title{Apply `stats::na.contiguous()` and pad result to original length.}
\usage{
ts_pad_contig(x)
}
\arguments{
\item{x}{Either a numeric vector, or `stats::ts()` object.}
}
\value{
An object of `class(x)` and `length(x)` with `NA`s, except for the
longest contiguous stretch in `x`
}
\description{
Apply `stats::na.contiguous()` to a timeseries `x` and pad the result back to
its original length with `NA`s. Designed for use in
`data.table[, j = pad_contiguous(x), by = g]`, where the replacement needs to
have the same length as the original.
}
\details{
Preserves `start` and `frequency` of `x`, if it is a `stats::ts()` object.
}
\examples{
x <- c(NA, 1, NA, NA, 1:5, NA)
ts_pad_contig(x)
x <- ts(x, start = c(2000, 1), frequency = 12)
ts_pad_contig(x)
}
