% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_math_letters.R
\name{latex_math_letters}
\alias{latex_math_letters}
\title{Toggle LaTeX math mode for leading letters.}
\usage{
latex_math_letters(string, convert = NULL, append = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{convert}{A character vector of hyphenated words to wrap.}

\item{append}{Boolean, should the words in \code{convert} be appended to the
implemented words.}
}
\value{
\code{string} with leading letters wrapped in \code{$}.
}
\description{
Wrap the leading letters in \emph{t}-test, \emph{F}-statistic,
\emph{p}-value, etc. in $ to toggle LaTeX math mode. Any combination to
leading letter and trailing word will also be modified (i.e. if \emph{t}-test
and \emph{F}-statistic are given, \emph{t}-statistic and \emph{F}-test will
also be matched).
}
\examples{
latex_math_letters("The t-test and z-score.", "z-score")
}
