% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_pad_contig.R
\name{ts_pad_contig}
\alias{ts_pad_contig}
\title{Apply \code{stats::na.contiguous()} and pad result to original length.}
\usage{
ts_pad_contig(x)
}
\arguments{
\item{x}{Either a numeric vector, or \code{stats::ts()} object.}
}
\value{
An object of \code{class(x)} and \code{length(x)} with \code{NA}s, except for the
longest contiguous stretch in \code{x}
}
\description{
Apply \code{stats::na.contiguous()} to a timeseries \code{x} and pad the result back to
its original length with \code{NA}s. Designed for use in
\code{data.table[, j = pad_contiguous(x), by = g]}, where the replacement needs to
have the same length as the original.
}
\details{
Preserves \code{start} and \code{frequency} of \code{x}, if it is a \code{stats::ts()} object.
}
\examples{
x <- c(NA, 1, NA, NA, 1:5, NA)
ts_pad_contig(x)
x <- ts(x, start = c(2000, 1), frequency = 12)
ts_pad_contig(x)
}
